+++ src/main/java/com/mycodefu/werekitten/level/data/Location.java
@@ -1,24 +1,24 @@
 package com.mycodefu.werekitten.level.data;
 
 public class Location{
-private int x;
-private int y;
-public int getX() {
+private double x;
+private double y;
+public double getX() {
 	return x;
 }
-public void setX(int x) {
+public void setX(double x) {
 	this.x = x;
 }
-public int getY() {
+public double getY() {
 	return y;
 }
-public void setY(int y) {
+public void setY(double y) {
 	this.y = y;
 }
 public Location() {
 
 }
-public Location(int x, int y) {
+public Location(double x, double y) {
 	setX(x);
 	setY(y);
 }
