+++ src/main/java/com/mycodefu/werekitten/level/data/Size.java
@@ -1,24 +1,24 @@
 package com.mycodefu.werekitten.level.data;
 
 public class Size {
-private int width;
-private int height;
-public int getWidth() {
+private double width;
+private double height;
+public double getWidth() {
 	return width;
 }
-public void setWidth(int width) {
+public void setWidth(double width) {
 	this.width = width;
 }
-public int getHeight() {
+public double getHeight() {
 	return height;
 }
-public void setHeight(int height) {
+public void setHeight(double height) {
 	this.height = height;
 }
 public Size() {
 
 }
-public Size(int width, int height) {
+public Size(double width, double height) {
 	setWidth(width);
 	setHeight(height);
 }
